refactor: change folders' architecture
