+++ src/endpoints/albums/dto/update-album.dto.ts
