+++ src/endpoints/albums/entities/album.entity.ts
