+++ src/endpoints/artists/artists.controller.ts
