+++ src/endpoints/artists/artists.module.ts
