+++ src/endpoints/artists/dto/create-artist.dto.ts
