+++ src/endpoints/artists/dto/update-artist.dto.ts
