+++ src/endpoints/artists/entities/artist.entity.ts
