+++ src/endpoints/favorites/dto/create-favorite.dto.ts
