+++ src/endpoints/favorites/dto/update-favorite.dto.ts
