+++ src/endpoints/favorites/entities/favorite.entity.ts
