+++ src/endpoints/favorites/favorites.module.ts
