+++ src/endpoints/users/entities/user.entity.ts
