+++ src/endpoints/users/users.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { IUser } from '../dataBase/dataBase.model';
-import { db } from '../dataBase/db';
+import { IUser } from '../../dataBase/dataBase.model';
+import { db } from '../../dataBase/db';
 
 @Injectable()
 export class UsersService {
